Add Dual Machines (indoor+outdoor)
A machine with an 'event_pair' and/or a 'stream_pair' field is a dual machine.
Both of these are objects with two fields: 'indoor' and 'outdoor'.
When the player is indoors (currently using the hardcoded "Indoor" and
"Outdoor" conditions that have to be defined in the json), the
streams/events in the 'indoor' field will be used, otherwise the 'outdoor'
field will be used.
If a dual machine has an 'event'/'stream' field but no 'event_pair'/'stream_pair'
field, 'event'/'stream' will be used by both machines.
This makes it possible to easily create machines with indoor and outdoor
versions.

Internally this is done by creating two machines.