Increase sound system streaming channels to 11 (from 4)
When travelling between different machines, the loops kept dying.
It seems that when a stream is played when all streaming channels
are already playing, one of them gets interrupted without any
notification, resulting in a zombie NoAttenuationMovingSound that
gets stuck thinking it's still playing while it's actually not.

This is just a quick fix, but one that should work for a while,
since it's unlikely I'd reach over 11 simultaneous streams.