Allow scanner modules to set values in their sheet
A scanner module's sheet now contains not only block counts, but
can also contain a number of values that are not block counts,
referred to as "variables". A variable must begin with a '.'
character, this sets it apart from a block name ('.' is an illegal
character in a block namespace, so it's impossible to have a conflict
with a block name this way).

The MediumScanner now sets the ".outsideness_score" in its sheet
instead of having its score asked by the getLastScore() method.

This change was required because I realized the MediumScanner wasn't
getting loaded outside of debug mode, since it was not required by my
json: only w_outsideness was. Although w_outsideness depends on
MediumScanner, MAtmos doesn't let modules depend on other modules
(only conditions and dynamics can depend on modules), so MediumScanner
wasn't recognized as a dependency, and wasn't loaded as a result.
With the new system, MediumScanner is directly referred to in the json,
so the problem is solved.