Allow overriding builtins
The reason I didn't previously is because I wanted to use the values
in their conditions to find the threshold above which scanning
should be stopped. This would have been very hard to do if they
could have been arbitrarily overriden.

However, I realized at this point I'm just defining constants.
So I did exactly that instead. The thresholds are localized
inside the scanners as constants, which the scanning algorithms
use to decide a boolean value.