Revert small scan height to 8, add weighted raycast scanner
I realized that the previous small scan size was a good choice:
it makes sense for it to be smaller vertically than horizontally,
since space feels more distant in the vertical direction than
the horizontal due to the existence of gravity.
For example, rooms are typially wider than they are tall,
so a scanner that shares these properties does a better job at
keeping sounds from other floors out of the scan.

However, it is sometimes useful to scan nearby objects further
than 4 blocks from the player vertically (see the river problem
that made me increase the scan height in the first place).
To make this possible, I made it possible for scanners to have
weighted versions (which increment the counter by a weight instead
of 1), and made the raycast scanner have a weighted version,
which gives a smaller weight the further away the block is.
This way it can be used to scan for nearby water while differentiating
from having a small body of water nearby, or a large body of water far
away (without weights, the same amount of rays could hit the water
in both cases).