Fix player Y pos being wrong
Sometime in Minecraft's history since 1.7.10, the player's
y coordinate was changed to mean the feet pos rather than the
eye pos. However, this wasn't updated in Matmos's code.

This resulted in abnormalities like can_rain_on turning false
because the player steps in water, or even just stands on a grass
path.