Add hint about restarting the sound system
When testing earlier in liteloader, occasionally the sound system
crashed, or sounds kept playing indefinitely.
The former of these is because of a bug in Paul's SoundSystem
which Forge fixes. I have not encountered the latter since switching
to Forge, but it may still be present.
It could be fixed by implementing an own system for starting and
stopping sounds, like how AmbientSounds did it.