First attempt at fixing streams failing to play
Occasionally, streams fail to play when they should. Reloading MAtmos
fixes this. This has been an issue for a while. But now, I have found
a sure way to recreate this: travelling between dimensions. If done
enough times, an ambience will fail to play.
The fix in this commit reduces the failure rate, and might fix the
bug in cases other than interdimensional travel. The reason it happens
in that case is due to a race condition that occurs: when changing
dimension, SoundManager.stopAllSounds gets called (caused by
SPacketRespawn to be more accurate). But sometimes, this happens
after SoundHelper.playStreaming gets called! MAtmos doesn't do any
checking to see if a sound has really started streaming. As a result,
those streams will not play until they are retriggered.