Make SoundHelper.startStreaming handle exception
Sometimes there was an IllegalArgumentException inside
SoundManager.playSound when travelling between dimensions
because the sound was already playing. But SoundManager.isSoundPlaying
was returning false, so this is a bug in Minecraft's code.
This is probably the same issue CreativeMD ran into here:
http://www.paulscode.com/forum/index.php?topic=14927.0
In my case, simply catching the exception should fix it.