Backport 1.7.10 branch
It works in a dev environment but there are a few small things left
to fix.

Note: LiteLoader builds were removed in this commit.
I don't use LiteLoader myself, and most people use Forge, so supporting
it is low priority for me. I built 1.12.2 for both LiteLoader and
Forge as a proof of concept for how modloader-agnostic MC-Commons is,
and because keeping LiteLoader support was little work.
But in this case I have decided not to bother with it.