Improve indoor scanning by raycast scanning through walls
Prior to this change, detecting if the player is indoors near the
surface was done by scan_air looking for nearby skylit blocks.
But blocks that are outdoors are not always skylit. For example, if
the player is inside a treehouse, it was detected as being deep
indoors, since the blocks outside are under leaves, so no nearby skylit
blocks were found.

To solve this, after a ray hits a surface, it is continued one block
beyond the surface. This effectively means one block thick walls and
ceilings are penetrated when searching for the sky.
By counting skylit blocks found this way, the raycast scanner calculates
its own ._is_near_surface judgement (._is_near_surface_own).
scan_air ors this with its own ._is_near_surface judgement (also renamed
to ._is_near_surface_own) to calculate the final ._is_near_surface.

Since rays are not stopped by leaves, this makes the previous example
case correctly be detected as the player being near the surface.