Build for Mixin 0.8 and 0.7
Based on my research:
The version of the mixin environment used by Minecraft is determined
by the mixin version of the first mixin-using mod that is loaded.
If a mod requires a mixin version higher than the mixin environment
that's already loaded, a crash occurs, and the game fails to load.
So, for example, suppose only Mod A and Mod B use mixins, and Mod A
is loaded before mod B.

* If Mod A uses Mixin 0.8 and Mod B uses Mixin 0.7, no crash occurs
  (unless Mod B tries to directly access Mixin internals, like
   what VanillaFix does)
* If Mod A uses Mixin 0.7 and Mod B uses Mixin 0.8, a crash occurs,
  and the game fails to load.

(Note: "Mod" in this context can refer to LiteLoader, which uses
 Mixin 0.8 at the time of writing. Since LiteLoader is loaded before
 any Forge mods, enabling LiteLoader can actually improve the
 compatibility of Forge mods!)

For this reason, updating this mod to Mixin 0.8 can actually cause
the game to stop functioning if another mod using Mixin 0.7 is loaded
earlier.

To provide compatibility with such mods, using Mixin 0.7 is required.
But for compatibility with up-to-date mods, using Mixin 0.8 is
required.

To this end, the 'forge' subproject has been split into a 'forge'
subproject which uses Mixin 0.8, and a 'forge-mixin0.7' subproject,
which uses Mixin 0.7.