Use cleaner method of overriding Sound System: transformers
Now that I have a better understanding of the way class loading
works, I realized that using a transformer is a much
simpler and cleaner way to override the Sound System
than the previous hacky method where I replaced the class loader.

The previous way worked incorrectly with other mods that used
ASM to modify the Sound System, such as Dynamic Surroundings.
Since I preloaded all the classes of the Sound System,
the transformers of other mods didn't get a chance to be
registered by the time class loading happened.

The new method inserts a transformer as the first one, which
simply changes the bytecode of Sound System classes to what's
read from the classes embedded in the MAtmos jar.