Let multiple machines control the volume of the same stream
Prior to this change, each machine with a stream had its own instance of the
stream, even if the streams were the same sound.
This was changed so that the two machines get a handle to the same stream.
This way, the volume of a stream can be dynamically changed.

This makes it impossible to play multiple instances of the same stream
at the same time, but I don't think that's ever desirable.

The interface of the SoundRelay was changed so that streams are no longer
started and stopped directly: this is now achieved by changing their volume.