Clean up properly when resource packs change
I don't know what the purpose of the brutal interruption was. I assume
it was meant to be an optimization. But it caused block change
listeners to not get removed when a sound pack was removed, which
caused duplicated sounds if the same soundpack was removed then readded.
That problem is no longer present, and I haven't found any new bugs
as a result of this change, so I'm going with it.