Disable liteloader builds
I have decided to stop supporting LiteLoader, as it has become
apparent that it's extra work that benefits few.
However, the codebase still references the mod loader through an
abstraction layer, so support for LiteLoader (or possibly even other
mod loaders) could be easily implemented.
Anyone is free to do so via a pull request.