Tick non-scanner modules on every tick
This makes them more reliable. For example, when digging with a sword,
default2016 would either play a swing sound on every swing or not, depending on
whether you started digging on an odd or even tick.

No impact on performance was found. And this doesn't seem to break any sound
packs, since the only time they check the magnitude of a delta value is when
they check for jumps between absolute values, which are instantaneous (and
therefore unaffected by tick rate.)